add state to handler return value, fix types around handlers

+++ mgmt/src/types.rs
@@ -1,4 +1,4 @@
 /// Basic quantities.
-pub type Time = u64;
-pub type Amount = u64;
+pub type Time    = u64;
+pub type Amount  = u128;
 pub type Address = cosmwasm_std::CanonicalAddr;
